LayoutTests/editing/spelling/text-replacement-after-typing-to-word-expected.txt

Before pressing enter:
PASS editor.textContent is 'YT?'
PASS successfullyParsed is true

TEST COMPLETE
You there?
